Backend/requirements.txt
fastapi
uvicorn
resend
python-dotenv
python-multipart